requirements.txt
appdirs==1.4.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.1.1
coverage==5.0.4
cssselect2==0.3.0
ddt==1.3.1
defusedxml==0.6.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-expects-json==1.4.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
Flask-SocketIO==4.2.1
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.4.1
freezegun==0.3.15
future==0.18.2
gast==0.3.3
get==2019.4.13
gevent==1.4.0
gevent-websocket==0.10.1
gitdb==4.0.2
gitdb2==4.0.2
GitPython==3.1.0
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lzstring==1.0.4
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==8.2.0
Naked==0.1.31
nltk==3.4.5
oauthlib==3.1.0
opt-einsum==3.2.0
pbr==5.4.4
Pillow==7.0.0
pluggy==0.13.1
post==2019.4.13
protobuf==3.11.3
public==2019.4.13
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
pyrsistent==0.16.0
pytest==5.4.1
pytest-cov==2.8.1
python-dateutil==2.8.1
python-engineio==3.12.1
python-ldap==3.2.0
python-socketio==4.5.1
PyYAML==5.3.1
query-string==2019.4.13
request==2019.4.13
requests-oauthlib==1.3.0
rsa==4.0
smmap==3.0.1
smmap2==3.0.1
SQLAlchemy==1.3.15
status==0.2.5
stevedore==1.32.0
termcolor==1.1.0
tinycss2==1.0.2
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
zipp==3.1.0